Split up files for better code understanding
