+++ src/yaacs/__init__.py
@@ -1,0 +1,4 @@
+from yaacs.cli import main
+
+if __name__ == "__main__":
+    main()
